# Add dimension handling interface for ingestion and segment creation