… benchmarks/src/main/java/io/druid/benchmark/indexing/IndexIngestionBenchmark.java
… extensions-core/kafka-indexing-service/src/test/java/io/druid/indexing/kafka/KafkaIndexTaskTest.java
… processing/src/main/java/io/druid/segment/DimensionMerger.java
… processing/src/main/java/io/druid/segment/DimensionMergerLegacy.java
… processing/src/main/java/io/druid/segment/IndexIO.java
… processing/src/main/java/io/druid/segment/IndexMerger.java
… processing/src/main/java/io/druid/segment/IndexMergerV9.java
… processing/src/main/java/io/druid/segment/IndexableAdapter.java
… processing/src/main/java/io/druid/segment/QueryableIndex.java
… processing/src/main/java/io/druid/segment/QueryableIndexIndexableAdapter.java
… processing/src/main/java/io/druid/segment/Rowboat.java
… processing/src/main/java/io/druid/segment/RowboatFilteringIndexAdapter.java
… processing/src/main/java/io/druid/segment/SimpleQueryableIndex.java
… processing/src/main/java/io/druid/segment/StorageAdapter.java
… processing/src/main/java/io/druid/segment/StringDimensionHandler.java
… processing/src/main/java/io/druid/segment/StringDimensionIndexer.java
… processing/src/main/java/io/druid/segment/StringDimensionMergerLegacy.java
… processing/src/main/java/io/druid/segment/StringDimensionMergerV9.java
… processing/src/main/java/io/druid/segment/column/DictionaryEncodedColumn.java
… processing/src/main/java/io/druid/segment/column/SimpleDictionaryEncodedColumn.java
… processing/src/main/java/io/druid/segment/incremental/IncrementalIndex.java
… processing/src/main/java/io/druid/segment/incremental/IncrementalIndexAdapter.java
… processing/src/main/java/io/druid/segment/incremental/IncrementalIndexStorageAdapter.java
… processing/src/main/java/io/druid/segment/incremental/OffheapIncrementalIndex.java
… processing/src/main/java/io/druid/segment/incremental/OnheapIncrementalIndex.java
… processing/src/test/java/io/druid/query/groupby/GroupByQueryRunnerTest.java
… processing/src/test/java/io/druid/segment/IndexMergerTest.java
… processing/src/test/java/io/druid/segment/filter/RowboatTest.java
… processing/src/test/java/io/druid/segment/incremental/IncrementalIndexTest.java
… processing/src/test/java/io/druid/segment/incremental/OnheapIncrementalIndexBenchmark.java