Description
--------------
This PR is a trimmed down version of #2760 that introduces a generic dimension handling infrastructure used for Long/Float dimension handling. However, only String dimensions are currently supported, to keep the patch at a reviewable size. Actual support for Longs/Floats will come in a later PR.

At a high level, type-specific handling for dimensions in the IncrementalIndex, IndexMerger has been extracted into a new interface. There are 3 associated classes with the new interface:

DimensionHandler: Top level class, provides comparators and used to create the other 3 classes
DimensionIndexer: Handles logic related to in-memory ingestion of rows (IncrementalIndex)
DimensionMerger: Handles logic related to merging of segments (IndexMerger)

StringDimensionHandler contains the implementations of these classes for String dimensions.

Validation
--------------
As this PR significantly refactors the indexing and segment creation code, it has the following risks:
- Possible performance degradation
- Regression failures

To address these risks, the following validation strategies were used:

* Performance Degradation
Benchmarks measuring the performance of IncrementalIndex row adds, IncrementalIndex reads, and segment merge/persist indicate that this PR does not introduce performance degradations in affected areas.

* Regression failures
This PR passes the Druid unit tests and integration tests. The hadoop integration test (ITHadoopIndexTest) was run/verified manually.



-----

Benchmarks from a c4.8xlarge

patch
```
Benchmark                                      (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt       Score       Error  Units
IncrementalIndexReadBenchmark.read                 true            750000     basic  avgt   50  334866.077 ±  5572.233  us/op
IncrementalIndexReadBenchmark.read                false            750000     basic  avgt   50  377260.561 ± 29860.209  us/op
IncrementalIndexReadBenchmark.readWithFilters      true            750000     basic  avgt   50  279812.716 ±  3370.998  us/op
IncrementalIndexReadBenchmark.readWithFilters     false            750000     basic  avgt   50  326096.952 ±  2870.442  us/op

Benchmark                        (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt       Score      Error  Units
IndexIngestionBenchmark.addRows      true             75000     basic  avgt   50  384164.000 ± 1966.527  us/op
IndexIngestionBenchmark.addRows     false             75000     basic  avgt   50  399771.783 ± 2446.356  us/op

Benchmark                   ( numSegments)  (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt        Score        Error  Units
IndexMergeBenchmark.merge                5      true             75000     basic  avgt   50  5448759.483 ± 159695.715  us/op
IndexMergeBenchmark.merge                5     false             75000     basic  avgt   50  5302718.895 ± 178994.287  us/op
IndexMergeBenchmark.mergeV9              5      true             75000     basic  avgt   50  4533349.734 ± 316551.870  us/op
IndexMergeBenchmark.mergeV9              5     false             75000     basic  avgt   50  4460183.171 ± 287253.950  us/op

Benchmark                        (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt        Score        Error  Units
IndexPersistBenchmark.persist        true             75000     basic  avgt   50  1306846.208 ± 170797.750  us/op
IndexPersistBenchmark.persist       false             75000     basic  avgt   50  1347429.256 ± 189093.734  us/op
IndexPersistBenchmark.persistV9      true             75000     basic  avgt   50  1205900.253 ± 181631.734  us/op
IndexPersistBenchmark.persistV9     false             75000     basic  avgt   50  1210533.192 ± 149366.322  us/op
```

master
```
Benchmark                                      (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt       Score       Error  Units
IncrementalIndexReadBenchmark.read                 true            750000     basic  avgt   50  338904.783 ±  9734.948  us/op
IncrementalIndexReadBenchmark.read                false            750000     basic  avgt   50  404326.322 ± 34762.840  us/op
IncrementalIndexReadBenchmark.readWithFilters      true            750000     basic  avgt   50  281539.062 ±  4068.100  us/op
IncrementalIndexReadBenchmark.readWithFilters     false            750000     basic  avgt   50  341720.741 ±  5290.862  us/op

Benchmark                        (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt       Score      Error  Units
IndexIngestionBenchmark.addRows      true             75000     basic  avgt   50  386509.637 ± 2585.987  us/op
IndexIngestionBenchmark.addRows     false             75000     basic  avgt   50  399927.398 ± 2514.297  us/op

Benchmark                   ( numSegments)  (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt        Score        Error  Units
IndexMergeBenchmark.merge                5      true             75000     basic  avgt   50  6430053.684 ± 298164.558  us/op
IndexMergeBenchmark.merge                5     false             75000     basic  avgt   50  6461107.881 ± 356905.476  us/op
IndexMergeBenchmark.mergeV9              5      true             75000     basic  avgt   50  5003605.414 ± 265774.052  us/op
IndexMergeBenchmark.mergeV9              5     false             75000     basic  avgt   50  5147752.762 ± 330098.556  us/op

Benchmark                        (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt        Score        Error  Units
IndexPersistBenchmark.persist        true             75000     basic  avgt   50  1366495.093 ± 179354.995  us/op
IndexPersistBenchmark.persist       false             75000     basic  avgt   50  1414207.294 ± 208999.251  us/op
IndexPersistBenchmark.persistV9      true             75000     basic  avgt   50  1297002.648 ± 133262.718  us/op
IndexPersistBenchmark.persistV9     false             75000     basic  avgt   50  1343949.275 ± 167867.933  us/op
```