@fjy 2016-07-26
@jon-wei merge conflicts now