@jon-wei 2016-08-02
rebasing, making adjustments based on new filtering changes