@drcrallen 2016-08-04
@jon-wei added a discuss label. Please remove it when ready for next round of review/merge