@jon-wei 2016-08-08
@drcrallen thanks, I've re-opened this now, benchmarks have been updated