@gianm 2016-08-09
Sick, it looks like `IndexMergeBenchmark` accidentally got faster :)