@jon-wei 2016-08-09
@gianm I haven't benchmarked this change in isolation, but I believe the IndexMerger speedup is from difference in implementation for single value rows here:

https://github.com/druid-io/druid/pull/3217/files#diff-4cb77a6cbc8f0534054954700125f6afL252
https://github.com/druid-io/druid/pull/3217/files#diff-aa6b53e2cf54c0d4132f9f1ead3f80c2R151

i.e., skipping the unneeded creation of an ArrayBasedIndexedInts and just making the single-value array directly