@jon-wei 2016-08-13
latest rebase changed the incrementalindex reading path slightly due to rebasing on top of #3344, re-ran IncrementalIndexReadBenchmark and performance hasn't changed:

```
Benchmark                                      (rollup)  (rowsPerSegment)  (schema)  Mode  Cnt       Score       Error  Units
IncrementalIndexReadBenchmark.read                 true            750000     basic  avgt   50  334866.077 ±  5572.233  us/op
IncrementalIndexReadBenchmark.read                false            750000     basic  avgt   50  377260.561 ± 29860.209  us/op
IncrementalIndexReadBenchmark.readWithFilters      true            750000     basic  avgt   50  279812.716 ±  3370.998  us/op
IncrementalIndexReadBenchmark.readWithFilters     false            750000     basic  avgt   50  326096.952 ±  2870.442  us/op
```