@jon-wei 2016-09-01
still addressing PR comments from @gianm, marking as WIP temporarily