@jon-wei 2016-09-01
Addressed PR comments from @gianm