@gianm 2016-09-07
Anyone else interested in reviewing? We do have two +1s already, but just checking. This has been open for a while and would like to get it in so we can begin building on top of it. If not then I'll merge it in a bit.