@gianm 2016-09-07
@drcrallen go for it; although please for @jon-wei's sake I'm hoping we can get this in before other things that might conflict with it (which is basically anything indexing related). He's already done countless rebases :)