@drcrallen 2016-09-07
if I can't get it done before EOD Friday I forfeit my right to complain about it