@jon-wei 2016-09-09
@drcrallen @gianm I moved the IndexIO type-specific validation into a method on DimensionHandler and updated parts of the javadocs in the new interfaces, can you take another look?