@drcrallen 2016-09-09
@jon-wei thanks.

I'm out of time for review, so I ask one last thing. Please check the output of indexing something like the wiki set or TPCH data, and report back on if they are binary equivalent pre and post this PR, and if they are not, make sure it is well understood why they are not.