@gianm 2016-09-09
👍  to the most recent changes. binary equivalence confirmed on wiki quickstart sample data:

```
18d4cf54a45961e30b264824e3757fbf912107fb  index/00000.smoosh
35406c92ad30b1d0cfa507fe9d9a48b14c4046ad  index/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  index/version.bin
18d4cf54a45961e30b264824e3757fbf912107fb  index2/00000.smoosh
35406c92ad30b1d0cfa507fe9d9a48b14c4046ad  index2/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  index2/version.bin
```