@drcrallen 2016-09-09
Cool, for binary equivalence and approval from at least two committers, please don't consider me blocking (but since I didn't do a full review I can't be one of the plus ones). I'm excited to see this move forward more.