@jon-wei 2016-09-10
Ran a binary equivalence check using IndexPersistBenchmark, equivalence is confirmed.

The legacy merger orders columns differently from the V9 merger, which is why the hashes differ.

Earlier I did catch an issue from this check, where this patch caused a fully null column to be erroneously written to the segment, this is fixed in a new commit, with a new test added for that case.

patch
```
legacy merger, with rollup
ed117be06cd41491dbfe50ce7acf2c35f436cbdb  1473468455019-0/00000.smoosh
124c1000535ad7b47c0eb44ea4f5ea941d51e72f  1473468455019-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468455019-0/version.bin

legacy merger, no rollup
ab59235f327ea608a8011d6c23b2b54c486581bd  1473468461807-0/00000.smoosh
884728fdad91189da23609ab414a9990bb940979  1473468461807-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468461807-0/version.bin

v9 merger, with rollup
f4209f50539e3ed5ca59d5d25b476802b8695275  1473468468730-0/00000.smoosh
bf54f4a77db631d8eb6dc59068462b26e01e3784  1473468468730-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468468730-0/version.bin

v9 merger, no rollup
7006cc8e01465603e1186fb99ad5023dbfdadaad  1473468475225-0/00000.smoosh
418b21c8ce2bb5b6dbaf3406894bab398c0d9c17  1473468475225-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468475225-0/version.bin
```

master
```
legacy merger, with rollup
ed117be06cd41491dbfe50ce7acf2c35f436cbdb  1473468651547-0/00000.smoosh
124c1000535ad7b47c0eb44ea4f5ea941d51e72f  1473468651547-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468651547-0/version.bin

legacy merger, no rollup
ab59235f327ea608a8011d6c23b2b54c486581bd  1473468658437-0/00000.smoosh
884728fdad91189da23609ab414a9990bb940979  1473468658437-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468658437-0/version.bin

v9 merger, with rollup
f4209f50539e3ed5ca59d5d25b476802b8695275  1473468665298-0/00000.smoosh
bf54f4a77db631d8eb6dc59068462b26e01e3784  1473468665298-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468665298-0/version.bin

v9 merger, no rollup
7006cc8e01465603e1186fb99ad5023dbfdadaad  1473468671867-0/00000.smoosh
418b21c8ce2bb5b6dbaf3406894bab398c0d9c17  1473468671867-0/meta.smoosh
febea93b2c7d75eb566bf43de2488821733303f5  1473468671867-0/version.bin
```