@gianm 2016-09-12
@jon-wei, looks like you used some jdk8-isms in a test file:

```
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-compiler-plugin:3.2:testCompile (default-testCompile) on project druid-processing: Compilation failure: Compilation failure:
[ERROR] /home/travis/build/druid-io/druid/processing/src/test/java/io/druid/segment/IndexMergerTest.java:[2021,24] cannot find symbol
[ERROR] symbol:   method sort(<nulltype>)
[ERROR] location: variable expectedColumnNames of type java.util.List<java.lang.String>
[ERROR] /home/travis/build/druid-io/druid/processing/src/test/java/io/druid/segment/IndexMergerTest.java:[2022,22] cannot find symbol
[ERROR] symbol:   method sort(<nulltype>)
[ERROR] location: variable actualColumnNames of type java.util.List<java.lang.String>
[ERROR] /home/travis/build/druid-io/druid/processing/src/test/java/io/druid/segment/IndexMergerTest.java:[2028,22] cannot find symbol
[ERROR] symbol:   method sort(<nulltype>)
[ERROR] location: variable expectedFilenames of type java.util.List<java.lang.String>
[ERROR] /home/travis/build/druid-io/druid/processing/src/test/java/io/druid/segment/IndexMergerTest.java:[2029,20] cannot find symbol
[ERROR] symbol:   method sort(<nulltype>)
[ERROR] location: variable actualFilenames of type java.util.List<java.lang.String>
[ERROR] -> [Help 1]
```